refactor: split src/meta.c into src/meta/ modules (#68)
src/meta.c held the metaclass end to end: the type object, the base
queries, the namespace build, and the class build/settle machinery.
This splits it along those seams, moving the code verbatim:

  src/meta/meta.c       the type object, entry points, GC, find_member
  src/meta/bases.c      base queries (equality, options, weakref slots)
  src/meta/namespace.c  namespace build, option application, refusals
  src/meta/class.c      build_struct_class and the settle machinery
  src/meta/meta.h       the shared types and the cross-file API

Each function keeps its old linkage: file-local helpers stay static in
their file (each file carries the forward declarations it needs, taken
from the old decl block), and only the 16 functions that cross a
boundary are extern, declared in meta.h. The header spells PyMemberDef
as `struct PyMemberDef` because it is read before types.h brings in
structmember.h on pre-3.12 Pythons, and the TESTING include gains a
`../` for the same relative-resolution rule.

No behaviour change: the matrix gate (3.10-3.15, free-threaded, the
in-file C tests, the analyzers) passes on the split as on the original.
