+++ build_config.py
@@ -14,7 +14,10 @@ BUILD: Final = BuildConfig(
         "src/construct.c",
         "src/fields.c",
         "src/hash.c",
-        "src/meta.c",
+        "src/meta/meta.c",
+        "src/meta/bases.c",
+        "src/meta/namespace.c",
+        "src/meta/class.c",
         "src/mixin.c",
         "src/options.c",
         "src/repr.c",
