+++ src/meta/class.c
@@ -1,0 +1,759 @@
+#include <Python.h>
+#include <stdbool.h>
+#include <stddef.h>
+#include <string.h>
+
+#include "../construct.h"
+#include "../fields.h"
+#include "meta.h"
+#include "../mixin.h"
+#include "../options.h"
+#include "../owned.h"
+#include "../result.h"
+#include "../types.h"
+
+static StructType * create_class(
+	PyTypeObject * metatype,
+	PyObject * name,
+	PyObject * bases,
+	PyObject * namespace
+);
+static PyTypeObject * winning_metatype(PyTypeObject * requested, PyObject * bases);
+static enum result install_fields(
+	StructType * struct_class,
+	StructType const * base,
+	struct field_plan const * plan,
+	struct options options,
+	bool resolves_body_eq
+);
+static enum result settle_planned(
+	StructType * struct_class,
+	StructType const * base,
+	PyObject * bases,
+	PyObject * name,
+	struct field_plan const * plan,
+	PyObject * original_namespace,
+	struct options options,
+	struct options inherited,
+	bool frozen_across_bases,
+	bool body_defines_eq,
+	bool inherits_body_eq,
+	bool derive_not_equal
+);
+static enum result settle_rebind(
+	StructType * struct_class,
+	PyObject * original_namespace,
+	char const * const * names,
+	bool from_mixin
+);
+static enum result restore_stripped(
+	StructType * struct_class,
+	PyObject * original_namespace,
+	PyObject * class_dict,
+	char const * const * names
+);
+static enum result refuse_unplanned(StructType const * struct_class);
+static enum result install_post_init(StructType * struct_class);
+static bool defines_own_init(StructType const * struct_class);
+static Py_ssize_t * resolve_slot_offsets(
+	StructType * struct_class,
+	StructType const * base,
+	PyObject * new_names,
+	Py_ssize_t field_count
+);
+static Py_ssize_t * resolve_member_offsets(
+	StructType * struct_class,
+	Py_ssize_t const * slot_offsets,
+	Py_ssize_t field_count,
+	Py_ssize_t * member_count
+);
+
+PyObject * build_struct_class(
+	PyTypeObject * const metatype,
+	StructType const * const base,
+	PyObject * const name,
+	PyObject * const bases,
+	PyObject * const original_namespace,
+	PyObject * const keywords
+) {
+	StructType const * const behaviour = find_behaviour_base(bases);
+	struct options const inherited = inherited_options(bases, behaviour);
+	struct options_request const request =
+		options_read(keywords, inherited, base != NULL && base->struct_field_count > 0);
+
+	if (request.tag == OPTIONS_REJECTED) {
+		return NULL;
+	}
+
+	if (
+		metatype == &StructMeta_Type &&
+		weakref_slot_is_new(request.options, bases) &&
+		winning_metatype(metatype, bases)->tp_new != StructMeta_new
+	) {
+		PyErr_SetString(
+			PyExc_TypeError,
+			"weakref=True cannot cross a metaclass __new__ that hands the build "
+			"off: the re-entered call cannot add the weakref slot"
+		);
+
+		return NULL;
+	}
+
+	struct field_plan plan = field_plan_build(base, original_namespace);
+
+	if (field_plan_failed(&plan)) {
+		return NULL;
+	}
+
+	if (
+		refuse_colliding_methods(original_namespace, plan.all_names, name) != RESULT_OK ||
+		refuse_mixin_method_fields(plan.all_names) != RESULT_OK ||
+		refuse_displaced_slots(
+				original_namespace,
+				plan.all_names,
+				weakref_expected(request.options, bases)
+			) !=
+			RESULT_OK
+	) {
+		field_plan_clear(&plan);
+
+		return NULL;
+	}
+
+	bool const body_defines_eq = PyDict_GetItemString(original_namespace, "__eq__") != NULL;
+
+	struct equality_source const inherited_equality = (
+		request.options.eq == inherited.eq && !body_defines_eq ? resolves_body_equality(bases) :
+		(struct equality_source){.tag = EQUALITY_RESOLVED, .from_a_body = false}
+	);
+
+	if (inherited_equality.tag == EQUALITY_FAILED) {
+		field_plan_clear(&plan);
+
+		return NULL;
+	}
+
+	bool const inherits_body_eq = inherited_equality.from_a_body;
+	bool const frozen_across_bases = request.options.frozen && any_struct_base_is_mutable(bases);
+
+	PY_OWNED(
+		namespace,
+		build_class_namespace(
+			original_namespace,
+			plan.all_names,
+			plan.new_names,
+			request.options,
+			base,
+			inherited,
+			frozen_across_bases,
+			body_defines_eq,
+			inherits_body_eq,
+			inherited_equality.needs_derived_not_equal
+		)
+	);
+	StructType * struct_class = (
+		namespace != NULL ? create_class(metatype, name, bases, namespace) :
+		NULL
+	);
+
+	if (struct_class != NULL) {
+		enum result const settled = (
+			struct_class->struct_field_names == NULL ? install_fields(
+				struct_class,
+				base,
+				&plan,
+				request.options,
+				body_defines_eq || inherits_body_eq
+			) :
+			settle_planned(
+				struct_class,
+				base,
+				bases,
+				name,
+				&plan,
+				original_namespace,
+				request.options,
+				inherited,
+				frozen_across_bases,
+				body_defines_eq,
+				inherits_body_eq,
+				inherited_equality.needs_derived_not_equal
+			)
+		);
+
+		if (settled != RESULT_OK) {
+			Py_CLEAR(struct_class);
+		}
+	}
+
+	field_plan_clear(&plan);
+
+	return (PyObject *) struct_class;
+}
+
+static StructType * create_class(
+	PyTypeObject * const metatype,
+	PyObject * const name,
+	PyObject * const bases,
+	PyObject * const namespace
+) {
+	PY_OWNED(type_args, PyTuple_Pack(3, name, bases, namespace));
+
+	if (type_args == NULL) {
+		return NULL;
+	}
+
+	PyTypeObject * const winner = winning_metatype(metatype, bases);
+	PyTypeObject * const builder = winner->tp_new == StructMeta_new ? winner : metatype;
+	PY_MOVABLE(created, PyType_Type.tp_new(builder, type_args, NULL));
+
+	if (created == NULL) {
+		return NULL;
+	}
+
+	if (!is_struct_class(created)) {
+		PyErr_Format(
+			PyExc_TypeError,
+			"%.200s.__new__ returned %.200s, which is not a struct class",
+			winner->tp_name,
+			Py_TYPE(created)->tp_name
+		);
+
+		return NULL;
+	}
+
+	return (StructType *) py_move(&created);
+}
+
+/* It is the third walk of `bases` in a class creation, after find_struct_base
+ * and _PyType_CalculateMetaclass. Measured, and smaller than the measurement:
+ * replacing the body with `return requested` leaves class creation at 9.77-9.87
+ * us for a 16-field class either way, so the walk is bounded by the width of
+ * that band rather than shown to be free. It is one Py_TYPE and one
+ * PyType_IsSubtype per base, and a class has one. */
+static PyTypeObject * winning_metatype(PyTypeObject * const requested, PyObject * const bases) {
+	PyTypeObject * winner = requested;
+
+	for (Py_ssize_t i = 0; i < PyTuple_GET_SIZE(bases); ++i) {
+		PyTypeObject * const candidate = Py_TYPE(PyTuple_GET_ITEM(bases, i));
+
+		if (PyType_IsSubtype(candidate, winner)) {
+			winner = candidate;
+		}
+	}
+
+	return winner;
+}
+
+static enum result settle_rebind(
+	StructType * const struct_class,
+	PyObject * const original_namespace,
+	char const * const * const names,
+	bool const from_mixin
+) {
+	PyObject * const source = (
+		from_mixin ? (PyObject *) &StructMixin_Type :
+		(PyObject *) &PyBaseObject_Type
+	);
+
+	for (char const * const * name = names; *name != NULL; name += 1) {
+		if (PyDict_GetItemString(original_namespace, *name) != NULL) {
+			continue;
+		}
+
+		PY_OWNED(bound, PyObject_GetAttrString(source, *name));
+		PY_OWNED(unicode_name, PyUnicode_FromString(*name));
+
+		if (
+			bound == NULL ||
+			unicode_name == NULL ||
+			PyType_Type.tp_setattro((PyObject *) struct_class, unicode_name, bound) < 0
+		) {
+			return RESULT_ERROR;
+		}
+	}
+
+	return RESULT_OK;
+}
+
+static enum result refuse_unplanned(StructType const * const struct_class) {
+	PyErr_Format(
+		PyExc_TypeError,
+		"%.200s.__new__ returned a struct class this call did not plan: its "
+		"metaclass is re-entered without the keywords or the class body's "
+		"defaults, so what it built is not what was asked for",
+		Py_TYPE(struct_class)->tp_name
+	);
+
+	return RESULT_ERROR;
+}
+
+static enum result settle_planned(
+	StructType * const struct_class,
+	StructType const * const base,
+	PyObject * const bases,
+	PyObject * const name,
+	struct field_plan const * const plan,
+	PyObject * const original_namespace,
+	struct options const options,
+	struct options const inherited,
+	bool const frozen_across_bases,
+	bool const body_defines_eq,
+	bool const inherits_body_eq,
+	bool const derive_not_equal
+) {
+	static char const * const comparison[] = {
+		"__eq__",
+		"__lt__",
+		"__le__",
+		"__gt__",
+		"__ge__",
+		NULL,
+	};
+	static char const * const not_equal[] = {"__ne__", NULL};
+	static char const * const representation[] = {"__repr__", NULL};
+	static char const * const mutability[] = {"__setattr__", "__delattr__", NULL};
+	static char const * const hash_name[] = {"__hash__", NULL};
+	static char const * const body_dunders[] = {
+		"__eq__",
+		"__ne__",
+		"__lt__",
+		"__le__",
+		"__gt__",
+		"__ge__",
+		"__repr__",
+		"__setattr__",
+		"__delattr__",
+		"__hash__",
+		"__init__",
+		"__post_init__",
+		"__new__",
+		NULL,
+	};
+	PY_OWNED(planned, PyList_AsTuple(plan->all_names));
+
+	if (planned == NULL) {
+		return RESULT_ERROR;
+	}
+
+	int const same_fields =
+		PyObject_RichCompareBool(struct_class->struct_field_names, planned, Py_EQ);
+
+	if (same_fields < 0) {
+		return RESULT_ERROR;
+	}
+
+	PY_OWNED(class_dict, struct_type_dict(&struct_class->heap_type.ht_type));
+
+	if (class_dict == NULL) {
+		return RESULT_ERROR;
+	}
+
+	PyObject * const built_name = (PyObject *) struct_class->heap_type.ht_name;
+	int const same_name = (
+		built_name != NULL ? PyObject_RichCompareBool(built_name, name, Py_EQ) :
+		0
+	);
+
+	if (same_name < 0) {
+		return RESULT_ERROR;
+	}
+
+	int const same_bases = PyObject_RichCompareBool(
+		((PyTypeObject *) struct_class)->tp_bases,
+		bases,
+		Py_EQ
+	);
+
+	if (same_bases < 0) {
+		return RESULT_ERROR;
+	}
+
+	if (
+		same_fields == 0 ||
+		same_name == 0 ||
+		same_bases == 0 ||
+		find_struct_base(((PyTypeObject *) struct_class)->tp_bases) != base
+	) {
+		return refuse_unplanned(struct_class);
+	}
+
+	bool const carries_slot = ((PyTypeObject *) struct_class)->tp_weaklistoffset != 0;
+
+	if (carries_slot != weakref_expected(options, bases)) {
+		return refuse_unplanned(struct_class);
+	}
+
+	struct binding_plan const bindings = binding_plan(
+		options,
+		inherited,
+		frozen_across_bases,
+		body_defines_eq,
+		inherits_body_eq,
+		derive_not_equal,
+		PyDict_GetItemString(original_namespace, "__hash__") != NULL
+	);
+
+	for (char const * const * name = body_dunders; *name != NULL; name += 1) {
+		if (
+			PyDict_GetItemString(class_dict, *name) != NULL &&
+			PyDict_GetItemString(original_namespace, *name) == NULL &&
+			!settled_by_the_plan(*name, bindings)
+		) {
+			return refuse_unplanned(struct_class);
+		}
+	}
+
+	if (restore_stripped(struct_class, original_namespace, class_dict, body_dunders) != RESULT_OK) {
+		return RESULT_ERROR;
+	}
+
+	if (
+		bindings.rebind_comparison &&
+		settle_rebind(struct_class, original_namespace, comparison, options.eq) != RESULT_OK
+	) {
+		return RESULT_ERROR;
+	}
+
+	if (
+		bindings.rebind_not_equal &&
+		settle_rebind(struct_class, original_namespace, not_equal, options.eq) != RESULT_OK
+	) {
+		return RESULT_ERROR;
+	}
+
+	if (bindings.answered_by_body && PyDict_GetItemString(original_namespace, "__ne__") == NULL) {
+		/* bind_not_equal's answered case on a live class: the fresh build
+		 * binds object's __ne__ when the body answered equality itself. */
+		PY_OWNED(object_ne, PyObject_GetAttrString((PyObject *) &PyBaseObject_Type, "__ne__"));
+		PY_OWNED(ne_name, PyUnicode_FromString("__ne__"));
+
+		if (
+			object_ne == NULL ||
+			ne_name == NULL ||
+			PyType_Type.tp_setattro((PyObject *) struct_class, ne_name, object_ne) < 0
+		) {
+			return RESULT_ERROR;
+		}
+	}
+
+	if (
+		bindings.rebind_representation &&
+		settle_rebind(struct_class, original_namespace, representation, options.repr) != RESULT_OK
+	) {
+		return RESULT_ERROR;
+	}
+
+	if (
+		bindings.rebind_mutability &&
+		settle_rebind(struct_class, original_namespace, mutability, options.frozen) != RESULT_OK
+	) {
+		return RESULT_ERROR;
+	}
+
+	switch (bindings.hash) {
+		case HASH_BODY_DEFINED:
+		case HASH_INHERITED_EQ:
+			break;
+		case HASH_NONE: {
+			PY_OWNED(hash_name_obj, PyUnicode_FromString("__hash__"));
+
+			if (
+				hash_name_obj == NULL ||
+				PyType_Type.tp_setattro((PyObject *) struct_class, hash_name_obj, Py_None) < 0
+			) {
+				return RESULT_ERROR;
+			}
+
+			break;
+		}
+		case HASH_BIND:
+			if (
+				settle_rebind(struct_class, original_namespace, hash_name, options.eq) !=
+				RESULT_OK
+			) {
+				return RESULT_ERROR;
+			}
+
+			break;
+	}
+
+	if (bindings.match_args_wanted) {
+		if (PyDict_SetItemString(class_dict, "__match_args__", planned) < 0) {
+			return RESULT_ERROR;
+		}
+	} else {
+		PyObject * const body_match_args = PyDict_GetItemString(
+			original_namespace,
+			"__match_args__"
+		);
+
+		if (body_match_args != NULL) {
+			if (PyDict_SetItemString(class_dict, "__match_args__", body_match_args) < 0) {
+				return RESULT_ERROR;
+			}
+		} else if (PyDict_DelItemString(class_dict, "__match_args__") < 0) {
+			if (PyErr_ExceptionMatches(PyExc_KeyError)) {
+				PyErr_Clear();
+			} else {
+				return RESULT_ERROR;
+			}
+		}
+	}
+
+	Py_SETREF(struct_class->struct_defaults, Py_NewRef(plan->defaults));
+	struct_class->struct_default_count = PyTuple_GET_SIZE(plan->defaults);
+	struct_class->struct_options = options;
+	struct_class->struct_resolves_body_eq = body_defines_eq || inherits_body_eq;
+
+	if (defines_own_init(struct_class)) {
+		struct_class->heap_type.ht_type.tp_new = Struct_new;
+		struct_class->heap_type.ht_type.tp_vectorcall = NULL;
+	} else {
+		struct_class->heap_type.ht_type.tp_vectorcall = Struct_vectorcall;
+	}
+
+	return install_post_init(struct_class);
+}
+
+static enum result restore_stripped(
+	StructType * const struct_class,
+	PyObject * const original_namespace,
+	PyObject * const class_dict,
+	char const * const * const names
+) {
+	for (char const * const * name = names; *name != NULL; name += 1) {
+		PyObject * const body_value = PyDict_GetItemString(original_namespace, *name);
+
+		if (body_value == NULL || PyDict_GetItemString(class_dict, *name) == body_value) {
+			continue;
+		}
+
+		PY_OWNED(unicode_name, PyUnicode_FromString(*name));
+
+		if (
+			unicode_name == NULL ||
+			PyType_Type.tp_setattro((PyObject *) struct_class, unicode_name, body_value) < 0
+		) {
+			return RESULT_ERROR;
+		}
+	}
+
+	return RESULT_OK;
+}
+
+static enum result install_fields(
+	StructType * const struct_class,
+	StructType const * const base,
+	struct field_plan const * const plan,
+	struct options const options,
+	bool const resolves_body_eq
+) {
+	PY_MOVABLE(field_names, PyList_AsTuple(plan->all_names));
+
+	if (field_names == NULL) {
+		return RESULT_ERROR;
+	}
+
+	Py_ssize_t const field_count = PyTuple_GET_SIZE(field_names);
+	Py_ssize_t * const offsets =
+		resolve_slot_offsets(struct_class, base, plan->new_names, field_count);
+
+	if (offsets == NULL) {
+		return RESULT_ERROR;
+	}
+
+	Py_ssize_t member_count = 0;
+	Py_ssize_t * const member_offsets =
+		resolve_member_offsets(struct_class, offsets, field_count, &member_count);
+
+	if (member_offsets == NULL) {
+		PyMem_Free(offsets);
+
+		return RESULT_ERROR;
+	}
+
+	struct_class->struct_field_names = py_move(&field_names);
+	struct_class->struct_defaults = Py_NewRef(plan->defaults);
+	struct_class->struct_slot_offsets = offsets;
+	struct_class->struct_member_offsets = member_offsets;
+	struct_class->struct_member_count = member_count;
+	struct_class->struct_field_count = field_count;
+	struct_class->struct_default_count = PyTuple_GET_SIZE(plan->defaults);
+	struct_class->struct_options = options;
+	struct_class->struct_resolves_body_eq = resolves_body_eq;
+
+	if (defines_own_init(struct_class)) {
+		struct_class->heap_type.ht_type.tp_new = Struct_new;
+	} else {
+		struct_class->heap_type.ht_type.tp_vectorcall = Struct_vectorcall;
+	}
+
+	return install_post_init(struct_class);
+}
+
+static bool defines_own_init(StructType const * const struct_class) {
+	return struct_class->heap_type.ht_type.tp_init != PyBaseObject_Type.tp_init;
+}
+
+static enum result install_post_init(StructType * const struct_class) {
+	PyObject * const hook = optional_attribute((PyObject *) struct_class, "__post_init__");
+
+	if (hook == NULL && PyErr_Occurred()) {
+		return RESULT_ERROR;
+	}
+
+	Py_XSETREF(struct_class->struct_post_init, hook);
+
+	return RESULT_OK;
+}
+
+static Py_ssize_t * resolve_slot_offsets(
+	StructType * const struct_class,
+	StructType const * const base,
+	PyObject * const new_names,
+	Py_ssize_t const field_count
+) {
+	Py_ssize_t * const offsets = PyMem_New(Py_ssize_t, field_count > 0 ? field_count : 1);
+
+	if (offsets == NULL) {
+		PyErr_NoMemory();
+
+		return NULL;
+	}
+
+	Py_ssize_t const inherited_count = base != NULL ? base->struct_field_count : 0;
+
+	for (Py_ssize_t i = 0; i < inherited_count; ++i) {
+		offsets[i] = base->struct_slot_offsets[i];
+	}
+
+	PyMemberDef const * const members = struct_heap_type_members(struct_class);
+	Py_ssize_t const member_count = Py_SIZE(struct_class);
+
+	for (Py_ssize_t i = 0; i < PyList_GET_SIZE(new_names); ++i) {
+		PyObject * const field_name = PyList_GET_ITEM(new_names, i);
+		struct member_lookup const found = find_member(members, member_count, field_name);
+
+		switch (found.tag) {
+			case MEMBER_LOOKUP_ERROR:
+				PyMem_Free(offsets);
+				return NULL;
+			case MEMBER_LOOKUP_MISSING:
+				PyErr_Format(PyExc_RuntimeError, "could not find slot offset for %R", field_name);
+				PyMem_Free(offsets);
+
+				return NULL;
+			case MEMBER_LOOKUP_FOUND:
+				offsets[inherited_count + i] = found.slot_offset;
+		}
+	}
+
+	return offsets;
+}
+
+static Py_ssize_t * resolve_member_offsets(
+	StructType * const struct_class,
+	Py_ssize_t const * const slot_offsets,
+	Py_ssize_t const field_count,
+	Py_ssize_t * const member_count
+) {
+	/* A non-struct base's own __slots__ members are not struct fields, so
+	 * the copy would never touch them without this table: one walk here, at
+	 * class creation, instead of rescanning every MRO dict on every copy.
+	 * The struct's own field descriptors are skipped by offset, and a
+	 * weakref slot is a getset descriptor, never a member one. */
+	*member_count = 0;
+
+	PyObject * const mro = struct_class->heap_type.ht_type.tp_mro;
+
+	for (Py_ssize_t i = 0; i < PyTuple_GET_SIZE(mro); ++i) {
+		PyTypeObject * const entry = (PyTypeObject *) PyTuple_GET_ITEM(mro, i);
+		PY_OWNED(entry_dict, struct_type_dict(entry));
+
+		if (entry_dict == NULL) {
+			return NULL;
+		}
+
+		Py_ssize_t position = 0;
+		PyObject * key;
+		PyObject * value;
+
+		while (PyDict_Next(entry_dict, &position, &key, &value)) {
+			if (!PyObject_TypeCheck(value, &PyMemberDescr_Type)) {
+				continue;
+			}
+
+			PyMemberDef const * const member = ((PyMemberDescrObject *) value)->d_member;
+
+			if (member == NULL || member->type != SLOT_MEMBER_TYPE) {
+				continue;
+			}
+
+			bool is_struct_field = false;
+
+			for (Py_ssize_t f = 0; f < field_count; ++f) {
+				if (member->offset == slot_offsets[f]) {
+					is_struct_field = true;
+					break;
+				}
+			}
+
+			if (!is_struct_field) {
+				++*member_count;
+			}
+		}
+	}
+
+	Py_ssize_t * const offsets = PyMem_New(Py_ssize_t, *member_count > 0 ? *member_count : 1);
+
+	if (offsets == NULL) {
+		PyErr_NoMemory();
+
+		return NULL;
+	}
+
+	Py_ssize_t written = 0;
+
+	for (Py_ssize_t i = 0; i < PyTuple_GET_SIZE(mro); ++i) {
+		PyTypeObject * const entry = (PyTypeObject *) PyTuple_GET_ITEM(mro, i);
+		PY_OWNED(entry_dict, struct_type_dict(entry));
+
+		if (entry_dict == NULL) {
+			PyMem_Free(offsets);
+
+			return NULL;
+		}
+
+		Py_ssize_t position = 0;
+		PyObject * key;
+		PyObject * value;
+
+		while (PyDict_Next(entry_dict, &position, &key, &value)) {
+			if (!PyObject_TypeCheck(value, &PyMemberDescr_Type)) {
+				continue;
+			}
+
+			PyMemberDef const * const member = ((PyMemberDescrObject *) value)->d_member;
+
+			if (member == NULL || member->type != SLOT_MEMBER_TYPE) {
+				continue;
+			}
+
+			bool is_struct_field = false;
+
+			for (Py_ssize_t f = 0; f < field_count; ++f) {
+				if (member->offset == slot_offsets[f]) {
+					is_struct_field = true;
+					break;
+				}
+			}
+
+			if (!is_struct_field) {
+				offsets[written++] = member->offset;
+			}
+		}
+	}
+
+	return offsets;
+}
